pyproject.toml
[project]
name = "skn21-3rd-1team"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "playwright>=1.57.0",
    "qdrant-client>=1.16.2",
    "ragas>=0.4.2",
    "rapidfuzz>=3.14.3",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-qdrant>=0.2.0",
    "langchain-huggingface>=0.1.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
    "einops>=0.8.1",
    "rank-bm25>=0.2.2",
    "flagembedding>=1.3.5",
]
